feat: moved fast mode to separate file and get_read_lengths_and_repeat_thresholds to the utils to not have circular imports
